// vitis-vision-front-end/src/app/core/api/models/vineyard-response.ts
/* tslint:disable */
/* eslint-disable */
export interface VineyardResponse {
  id?: number;
  companyName?: string;
  dbaName?: string;
  city?: string;
  email?: string;
  phoneNumber?: string;
  streetAddress?: string;
  zipCode?: string;
  createdAt?: string;
  createdBy?: string;
  lastUpdatedAt?: string;
  lastUpdatedBy?: string;
}